Code/cloud-native-handout-prithiraj-bhuyan/project3/dapr/worker/pyproject.toml
[project]
name = "worker"
version = "0.1.0"
description = "my llm server"
authors = [
    {name = "Prithiraj Bhuyan"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.125.0,<0.126.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "llama-cpp-python (>=0.3.16,<0.4.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "grpcio-tools (>=1.76.0,<2.0.0)",
    "dapr (>=1.16.0,<2.0.0)",
    "dapr-ext-fastapi (>=1.16.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "mypy (>=1.19.1,<2.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
    "coverage (>=7.13.0,<8.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "ruff (>=0.14.10,<0.15.0)",
    "pydoclint (>=0.8.3,<0.9.0)"
]
